pinecone-qa/requirements.txt
streamlit==1.25.0
langchain==0.0.268
openai==0.27.8
tiktoken==0.4.0
pypdf==3.15.1
pinecone-client==2.2.2
google-generativeai==0.3.2
